Jobs migration

+++ BackEnd/database/migrations/2024_05_14_112112_create_freelancing_job_competetors_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('individual_id')->constrained()->cascadeOnDelete();
             $table->foreignId('job_id')->constrained('freelancing_jobs')->cascadeOnDelete();
             $table->string("description");
-            $table->integer("salary");
+            $table->float("salary");
             $table->timestamps();
         });
     }
